reorganize js files
