+++ webroot/js/utils/helpers.js
@@ -1,22 +1,3 @@
-export const URL_STATUS = `/status`;
-export const URL_CHAT_HISTORY = `/chat`;
-export const URL_CUSTOM_EMOJIS = `/emoji`;
-export const URL_CONFIG = `/config`;
-
-// TODO: This directory is customizable in the config.  So we should expose this via the config API.
-export const URL_STREAM = `/hls/stream.m3u8`;
-export const URL_WEBSOCKET = `${location.protocol === 'https:' ? 'wss' : 'ws'}://${location.host}/entry`;
-
-export const TIMER_STATUS_UPDATE = 5000; // ms
-export const TIMER_DISABLE_CHAT_AFTER_OFFLINE = 5 * 60 * 1000; // 5 mins
-export const TIMER_STREAM_DURATION_COUNTER = 1000;
-export const TEMP_IMAGE = 'data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7';
-
-export const MESSAGE_OFFLINE = 'Stream is offline.';
-export const MESSAGE_ONLINE = 'Stream is online';
-
-export const URL_OWNCAST = 'https://github.com/gabek/owncast'; // used in footer
-
 export function getLocalStorage(key) {
   try {
     return localStorage.getItem(key);
