+++ webroot/standalone-chat.html
@@ -19,7 +19,7 @@
 
     <script type="module">
       import { render, html } from "https://unpkg.com/htm/preact/index.mjs?module";
-      import StandaloneChat from './js/chat/standalone.js';
+      import StandaloneChat from './js/standalone-chat-app.js';
 
       const messagesOnly = false;
